// src/app/models/photon-enums.ts
export enum PhotonFilters {
	None = "",
	Radio = "radio",
	LoFi = "lofi",
	Oceanic = "oceanic",
	Islands = "islands",
	Marine = "marine",
	Seagreen = "seagreen",
	Flagblue = "flagblue",
	Liquid = "liquid",
	Diamante = "diamante",
	Twenties = "twenties",
	Mauve = "mauve",
	Vintage = "vintage",
	Bluechrome = "bluechrome"
}

export enum PhotonEffects {
	None = "",
	Solarize = "solarize",
	Oil_Painting = "oil_painting",
	Offset_Red = "offset_red",
	Ryo = "ryo"
}

export enum PhotonTransform {
	None = "",
	Flip_Horizontal = "flip_horizontal",
	Flip_Vertical = "flip_vertical"
}